the .to16() method on registers now returns a 16 bit number, which revaled bugs in other opcodes

+++ Include/Registers.h
@@ -6,7 +6,7 @@
 //the z80 uses eightbit regiisters A, F, B, C, D, E, H ,L however they are used in pairs
 	struct sixReg {
 		uint8_t high, low;
-		uint8_t to16() {
+		uint16_t to16() {
 			uint16_t x = high << 8 | low;
 			return x;
 		};
